Modularized the greeting functions. Modularized the creating of the bot instance
