+++ bot.js
@@ -1,49 +1,21 @@
-const SlackBot = require('slackbots');
-const { token } = require('./private/token.js')
+const { bot } = require('./createBot.js')
+const { sayHelloOnUserArrival, sayHelloOnBotLogIn } = require('./greetings')
 
-// create a bot
-var bot = new SlackBot({
-    token: token, // Add a bot https://my.slack.com/services/new/bot and put the token
-    name: 'Johnny Five'
-});
 
 bot.on('start', function() {
-    // more information about additional params https://api.slack.com/methods/chat.postMessage
-    var params = {
-        icon_emoji: ':robot_face:'
-    };
-
     // define channel, where bot exist. You can adjust it there https://my.slack.com/services
 
-    bot.postMessageToChannel('general', 'Hey Humans!  I\m back!', params);
-
-    // define existing username instead of 'user_name'
-    // bot.postMessageToUser('westley', 'Hello Bozo, testing!', params);
-    // bot.postMessageToUser('jaredshane', 'Hello Bozo!', params);
-
-    // If you add a 'slackbot' property,
-    // you will post to another user's slackbot channel instead of a direct message
-    // bot.postMessageToUser('westley', 'meow!', { 'slackbot': true, icon_emoji: ':cat:' });
-
-    // define private group instead of 'private_group', where bot exist
-    // bot.postMessageToGroup('private_group', 'meow!', params);
-
-
-    // setTimeout(()=>{
-    //     console.log(bot.getUser('U4N0F05NV'))
-    //     // bot.getUser('westley')
-    // }, 4000)
+    // sayHelloOnBotLogIn()  // comment out during testing - its annoying
 });
 
 
 bot.on('message', (event) => {
-    // console.log('message event fired')
     console.log(event)
 
 
     // check if user is logging in to slack
     if(event.type === 'presence_change' && event.presence === 'active') {
-        sayHelloOnArrival(event.user)
+        sayHelloOnUserArrival(event.user)
     }
 
 
@@ -60,19 +32,6 @@ bot.on('message', (event) => {
     }
 
 })
-
-
-const sayHelloOnArrival = (userId) => {
-
-    // console.log("say hello func called")
-
-    let params = {
-        icon_emoji: ':robot_face:'
-    };
-
-    // use postMessage to send messages using user IDs
-    bot.postMessage(userId, 'Hello Bozo!', params);
-}
 
 
 const checkMessageForBotCommand = (channelText) => {
@@ -111,24 +70,9 @@ const checkMessageForBotCommand = (channelText) => {
 const processBotCommand = (command) => {
     console.log("process command test", command)
 
-    // // if (command.type !== 'say') {
-    // //     return false
-    // // }
-
     let params = {
         icon_emoji: ':robot_face:'
     };
-
-    // const commandList = {
-    //     "say": bot.postTo('general', command.argument, params),
-    //     "tictactoe": "something will happen here"
-    // }
-
-    // console.log("type right here", command.type)
-
-    // // console.log(commandList[command.type])
-    // return commandList[command.type] || console.log("Improper command entered")
-
 
     // SWITCH STATEMENT WAY THAT WORKS
 
@@ -138,8 +82,8 @@ const processBotCommand = (command) => {
             bot.postTo('general', command.argument, params);
             break;
 
-        case 'tic':
-            console.log("tactoe")
+        case 'killbot':
+            process.exit()
             break;
 
         default:
