+++ package.json
@@ -4,7 +4,8 @@
   "description": "Slack game - Tic Tac Toe",
   "main": "bot.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "nodemon ./bot.js"
   },
   "repository": {
     "type": "git",
